Move `MultiParentCasperTestUtil.copyStorage` et al to `Resources`
