+++ casper/src/test/scala/coop/rchain/casper/helper/RhoSpec.scala
@@ -3,7 +3,7 @@ package coop.rchain.casper.helper
 import cats.effect.{Concurrent, Sync}
 import cats.implicits._
 import coop.rchain.casper.MultiParentCasperTestUtil
-import coop.rchain.casper.MultiParentCasperTestUtil.copyStorage
+import coop.rchain.casper.util.rholang.Resources.copyStorage
 import coop.rchain.casper.genesis.contracts.TestUtil
 import coop.rchain.casper.genesis.contracts.TestUtil.eval
 import coop.rchain.casper.protocol.DeployData
