+++ casper/src/test/scala/coop/rchain/casper/util/rholang/Resources.scala
@@ -1,14 +1,16 @@
 package coop.rchain.casper.util.rholang
 
-import java.nio.file.Path
+import java.nio.file.StandardCopyOption.REPLACE_EXISTING
+import java.nio.file.{Files, Path}
 
-import cats.effect.{Concurrent, ContextShift, Resource}
+import cats.effect.{Concurrent, ContextShift, Resource, Sync}
 import cats.implicits._
 import cats.temp.par
 import coop.rchain.metrics
 import coop.rchain.metrics.{NoopSpan, Span}
 import coop.rchain.rholang.Resources.{mkRuntimeAt, mkTempDir}
 import coop.rchain.shared.Log
+import coop.rchain.shared.PathOps.RichPath
 import monix.eval.Task
 import monix.execution.Scheduler
 
@@ -35,4 +37,35 @@ object Resources {
     } yield runtimeManager
   }
 
+  case class StoragePaths(
+      blockStoreDir: Path,
+      blockDagDir: Path,
+      rspaceDir: Path
+  )
+
+  def copyStorage[F[_]: Sync](
+      storageTemplatePath: Path
+  ): Resource[F, StoragePaths] =
+    for {
+      storageDirectory <- Resource.make[F, Path](
+                           Sync[F].delay {
+                             val dir = Files.createTempDirectory(s"casper-test-")
+                             copyDir(storageTemplatePath, dir)
+                           }
+                         )(dir => Sync[F].delay { dir.recursivelyDelete() })
+      blockStoreDir = storageDirectory.resolve("block-store")
+      blockDagDir   = storageDirectory.resolve("block-dag-store")
+      rspaceDir     = storageDirectory.resolve("rspace")
+    } yield StoragePaths(
+      blockStoreDir = blockStoreDir,
+      blockDagDir = blockDagDir,
+      rspaceDir = rspaceDir
+    )
+
+  private def copyDir(src: Path, dest: Path): Path = {
+    Files
+      .walk(src)
+      .forEach(source => Files.copy(source, dest.resolve(src.relativize(source)), REPLACE_EXISTING))
+    dest
+  }
 }
